% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.flexsurv-package}
\alias{mlr3learners.flexsurv}
\alias{mlr3learners.flexsurv-package}
\title{mlr3learners.flexsurv: Learners from the {flexsurv} package for 'mlr3'}
\description{
Adds learners from the {flexsurv} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
